+++ sample-compose-fragments/src/main/java/uk/gov/hmrc/sample_compose_fragments/presentation/screens/atoms/ButtonScreen.kt
@@ -20,13 +20,13 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import uk.gov.hmrc.components.compose.atom.button.IconButton
 import uk.gov.hmrc.components.compose.atom.button.PrimaryButton
 import uk.gov.hmrc.components.compose.atom.button.SecondaryButton
-import uk.gov.hmrc.components.compose.ui.theme.HmrcTheme
 import uk.gov.hmrc.components.compose.ui.theme.HmrcTheme.dimensions
+import uk.gov.hmrc.components.compose.ui.theme.HmrcAllDevicePreview
 import uk.gov.hmrc.sample_compose_components.R
+import uk.gov.hmrc.sample_compose_fragments.presentation.screens.sampletemplate.HmrcPreview
 import uk.gov.hmrc.sample_compose_fragments.presentation.screens.sampletemplate.ScreenScrollViewColumn
 
 @Composable
@@ -46,10 +46,10 @@ fun ButtonScreen() {
     }
 }
 
-@Preview(showBackground = true)
+@HmrcAllDevicePreview()
 @Composable
 fun ButtonScreenPreview() {
-    HmrcTheme {
+    HmrcPreview {
         ButtonScreen()
     }
 }