+++ sample-compose-fragments/src/main/java/uk/gov/hmrc/sample_compose_fragments/presentation/screens/sampletemplate/ComposableTemplate.kt
@@ -79,3 +79,12 @@ fun ExamplesSlot(content: @Composable () -> Unit) {
     )
     content()
 }
+
+@Composable
+fun HmrcPreview(content: @Composable () -> Unit) {
+    HmrcTheme {
+        HmrcSurface(
+            content = content
+        )
+    }
+}